+++ docs/changes.rst
@@ -4,9 +4,11 @@ Changelog
 1.2.4 (2018/12)
 ---------------
 
-Fixes:
+Improvements:
 
 - ``Apyfal.host.Host`` now checks instance boot using port 22 instead of 80.
+- ``Apyfal.host.Host`` can now be used to start a custom instance instead of an
+  predefined accelerator instance more easily.
 
 1.2.3 (2018/10)
 ---------------